requirements.txt
scapy>=2.4.5
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.0
joblib>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tkinter